refactor: TicketMapper 파일 생성
