+++ app/src/main/java/woowacourse/movie/view/reservation/detail/ReservationPresent.kt
@@ -1,11 +1,11 @@
 package woowacourse.movie.view.reservation.detail
 
 import android.os.Bundle
+import woowacourse.movie.data.mapper.toUi
 import woowacourse.movie.domain.Movie
 import woowacourse.movie.domain.Ticket
 import woowacourse.movie.view.home.movies.toMovieUi
 import woowacourse.movie.view.home.theater.Showing
-import woowacourse.movie.view.reservation.toUi
 import java.time.LocalDateTime
 
 class ReservationPresent(
